Move tests into argument specific files.
